+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
-## 0.0.1
+## 1.1.0
+- add UI notification to make the intent of temporary mode more clear
+
+- add line in the readme about how pros and cons are implemented
+
+- rework Powers UI a little to try to make how Pros and Cons are implemented more clear
+
+- fix temporary outline for talents not appearing in temp. buffs mode
+
+- foundry 13 compatability changes: refactor hooks, fix activeeffect creation. this version should work with both 12 and 13
+
+- add Playwright and start adding test cases
+
+- improve speed of package change
+
+- fixed applying damage to minions
+
+
+## 1.0.0
 
 Initial release